Request for the formalization of the PUG token within carbonswap
We are the mafia that protects energy web chain for scammers and weak hands.. 5% of the pugs sent in any type of transaction are burned,or dead, as you like say it, streets are dangerous, pugs are shot in each transaction and unfortunately the casualties are 5%...We want to be the biggest and memera mafia of ewc and all the blockchains, faithful to clean energy, defending all the susus by patrolling carbonswap. 

We are a memera mafia, but we want to offer the maximum possible utility, we hope to offer NFTs to the top holders, and offer part of the money obtained to kennels and to charity.