Request for the formalization of the PUG token within Carbonswap Community List
Token that was conceived as a memetoken but has evolved over time as a utility token, PUG provides through our DEX (Decentralised Extension) an extra layer to carbonswap, through CLPs and tokens native to the PUG ecosystem you can enter farming pools, staking, lotteries, and more things that are being worked on such as NFTs.

We are the mafia that protects the energy chain from scammers and weak hands. 5% of the pugs that are sent in any kind of transaction are burned, or die, however you want to put it, the streets are dangerous, pugs are shot on every transaction and unfortunately the casualties are 5%...We want to be the biggest friendly mafia in ewc and all blockchains, true to clean energy, defending all susus patrolling carbonswap.

We have clean energy and helping as a fundamental pillar within the project, as will be seen.